Add to-do tasks Add more parameters to API requests based on API documentation
