+++ pixleesdk/src/main/java/com/pixlee/pixleesdk/data/repository/BasicRepository.java
@@ -48,7 +48,7 @@ public class BasicRepository implements BasicDataSource {
     @Override
     public Call<PhotoResult> postMedia(String api_key, JSONObject json) {
         if (PXLClient.secretKey == null ) {
-            throw new IllegalArgumentException("no apiKey, please set apiKey before start");
+            throw new IllegalArgumentException("no secretKey, please set secretKey before start");
         }
         String signiture = null;
         try {
